api commit

+++ app/Http/Controllers/UserController.php
@@ -28,6 +28,7 @@ class UserController extends Controller
     {
         $roles = Role::pluck('name','name')->all();
         
+       
         return view('role-permission.user.create', ['roles' => $roles]);
     }
 
@@ -53,6 +54,15 @@ class UserController extends Controller
                     
                     ['created_at' => now(), 'updated_at' => now()]
                     );
+
+                     // ✅ Log the create action (no old/new values)
+        DB::table('edit_logs')->insert([
+            'table_name' => 'users',
+            'record_id'  => $user->id,
+            'action'     => 'created',
+            'updated_by' => Auth::id(),
+            'created_at' => now(),
+        ]);
 
         return redirect('/users')->with('success','User created successfully with roles');
     }
@@ -107,14 +117,50 @@ class UserController extends Controller
                 'updated_at' => now(),
             ]);
         }
+
+        $oldData = $user->replicate();
+        // ✅ Compare old vs new and log only changed fields
+        $changedFields = $user->getChanges(); // only changed attributes
+
+        foreach ($changedFields as $field => $newValue) {
+            if (in_array($field, ['updated_at'])) continue; // skip timestamps
+
+            $oldValue = $oldData->$field ?? null;
+
+            DB::table('edit_logs')->insert([
+                'table_name' => 'users',
+                'record_id' => $user->id,
+                'field' => $field,
+                'old_value' => $oldValue,
+                'new_value' => $newValue,
+                'action' => 'updated',
+                'updated_by' => Auth::id(),
+                'created_at' => now(),
+            ]);
+        }
         return redirect('/users')->with('success','User Updated Successfully with roles');
     }
 
     public function destroy($userId)
     {
+        $isDReferenced = DB::table('distributors')->where('user_id', $userId)->exists();
+        $isEReferenced = DB::table('employees')->where('created_by', $userId)->exists();
+        if ($isDReferenced || $isEReferenced) {
+            return redirect()->back()->with('failure', 'User cannot be deleted because it is referenced in another table.');
+        }
         $user = User::findOrFail($userId);
+        $user->is_deleted=1;
         $user->delete();
+        
 
+        // ✅ Log the delete action only (no old/new value)
+        DB::table('edit_logs')->insert([
+            'table_name' => 'users',
+            'record_id' => $user->id,
+            'action' => 'deleted',
+            'updated_by' => Auth::id(),
+            'created_at' => now(),
+        ]);
         return redirect('/users')->with('success','User Delete Successfully');
     }
 }
